v4l2-tracer: Rewrite libv4l2tracer.so lookup
The lookup had become messy and unreliable. In this rewrite, we
honor LD_PRELOAD if already set. Following, we use the new env
V4L2_TRACER if set. This env is now set by meson devenv to point
to the build version whenever uninstalled setup is used. Finally,
lookup for the installed path. Simply fail without fallback if the
env are set but aren't pointing to a file.

Signed-off-by: Nicolas Dufresne <nicolas.dufresne@collabora.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>