cec-compliance: add --show-timestamp option
Add a new option to prefix the test name with a timestamp. This
helps associating the test commands with the timestamps you see when
monitoring the CEC traffic. Without this it is hard to discover
when each test starts and ends and which CEC messages are actually
from each specific test.

Hopefully this will help with that.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>