v4l-utils: (c)v4l-helpers.h: support more than 32 buffers
The v4l2-compliance test failed (segfault) when used with the
vivid driver which supports more than 32 buffers. It turned
out that the helper headers did not allocate enough space for
the buffer information when there are more than 32 buffers.

Rework the code so that it correctly queries the max number
of buffers and allocates space for it.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>