cec-ctl: avoid confusing error message
In the --stress-test-random-standby-wakeup-cycle test it attempted
to transmit <Standby> or <Image View On> for up to 50 times with a
0.2s wait in between each attempt. If all attempts fail, then log
the transmit error.

However, a 10 second wait is overkill here, retrying it 10 times
(2 seconds) is sufficient. And if the error is ENONET (i.e. the HPD
of the HDMI receiver is low), then don't log the error message.

This is a valid situation in this test if the HPD happened to be low
due to an earlier Standby and if it didn't see/honor the Image View
On that followed. So seeing a "Machine is not on the network" error
is confusing.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>