v4l2-ctl: --list-devices --verbose will list current input/output
It can be handy to see what the current input or output name is of
a given v4l2 device: that often gives useful information to determine
which device to use. So if --verbose is added to --list-devices, then
it will report that information in parenthesis after the device name.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>