v4l2-compliance: do not warn for unexpected vivid_ro_ctrl values
The vivid driver does not produce reliable sequence counters,
the first sequence counter may be > 0, and there are gaps in the
counters when there shouldn't be.

Until this is fixed (work in progress) replace the warning when
an unexpected vivid_ro_ctrl value is seen by an info message.

Once vivid is fixed this can be reworked.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>