utils: media-ctl: Support accessing the subdev TRY state
Add a -W/--which argument to media-ctl to select which state to operate
on. Default to the ACTIVE state to preserve the current behaviour.

Despite the fact that all values set on the TRY state are lost when
media-ctl terminates, support for the TRY state is useful in order to
retrieve the default configuration of subdevs, or to try configurations.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>
Acked-by: Sakari Ailus <sakari.ailus@linux.intel.com>