qvidcap: fix core dump
The 'cv4l_queue q' variable went out of scope too early,
causing a core dump when g_dataptr was called later.

I suspect it was pure luck that this ever worked, but the changes
to support more than 32 buffers in combination with vivid supporting
this new feature, caused this bug to surface.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>