qv4l2: always set m_haveFramebufferSRGB to false
With Qt6 glEnable(GL_FRAMEBUFFER_SRGB) results in dark colors for some
reason. It is fine with Qt5. Since qvidcap also ignores this (i.e. it never
enables sRGB) I decided to do the same for qv4l2, so just set this to
false.

At some point qv4l2 should really switch to the qvidcap openGL
code since it is getting to be quite a mess.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>