contrib/test/test-media: re-add cec-compliance -A tests
In the past the cec-compliance -A tests (i.e. testing the adapter
itself) was disabled since vivid did not handle the Lost Arbitration
situation correctly, so it would always fail.

However, vivid has been fixed quite a long time ago back in September
2021, but the test-media script was never updated to re-enable this
adapter compliance test.

So add it back as it is safe to do so now.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>