v4l2-compliance: only do request-queued checks while streaming is off
The checks that verify if the queued request (and associated buffer) is
actually in the queued state should only be done while streamon hasn't
been called. Otherwise the request might already been processed and is
no longer in the queued state, which causes hard to reproduce failures.

This also reinstates the V4L2_BUF_FLAG_QUEUED check as that failed for
the same reason. That test is fine as long as streamon wasn't called.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Fixes: 37d02fb3e909 ("v4l2-compliance: testRequests(): drop V4L2_BUF_FLAG_QUEUED check")