v4l2-ctl: add --list-devices-input/output options
With these options you can filter the listed devices by
the input or output name of the currently selected input
or output of the device.

Also move the handling of the list-devices options to the
beginning and exit right after. This fixes an issue if there
is no /dev/video0 (e.g. if the first device is /dev/video2),
and v4l2-ctl --list-devices would fail with an error in that
case (can't open video0).

Signed-off-by: Hans Verkuil <hverkuil@xs4all.nl>