qv4l2/qvidcap: bind shader program in paintGL()
Starting in Qt 6.7.0, vertex buffers and shader programs are unbound
just before calling QOpenGLWidget::paintGL(). In the case of qv4l2
and qvidcap this means that in order to stay compatible with Qt 6.7
the shader program has to be bound in paintGL each time, otherwise
nothing would be rendered.

In the case of qv4l2 it would still render (although I am not quite
sure how), but it would generate openGL errors, and in the case of
qvidcap nothing would be shown after the first frame was rendered.

Adding the extra bind calls fixes this issue.

Many thanks to Laurent Pinchart for pointing me to a similar fix in
libcamera.

Signed-off-by: Hans Verkuil <hverkuil@xs4all.nl>