edid-decode: move edid-decode into v4l-utils
The edid-decode utility was maintained in its own git repository:

	git://linuxtv.org/hverkuil/v4l-utils.git

This patch moves it into the v4l-utils git repository.

This helps with versioning (v4l-utils has regular releases), and
it makes it easy to check if the EDIDs included in v4l2-ctl are
compliant. In addition, I plan to add a new script that tests if
displays are CEC and EDID compliant, and for that it makes sense
if edid-decode is included here.

Signed-off-by: Hans Verkuil <hverkuil@xs4all.nl>