cec-ctl/edid-decode: use seconds for the reliability test
The cec-ctl and edid-decode reliability tests both use 'count'
as a measure of how long these tests should take. But duration
in seconds is a much more useful number, especially since you
typically want to run both tests at the same time for the same
HDMI cable for the same duration.

So change the argument to be the number of seconds that the
test must run. '0' still means 'forever'.

Signed-off-by: Hans Verkuil <hverkuil@xs4all.nl>