test_display.py: add log options, checked with linter
Renamed to test_display.py ('-' in names is not recommended in
python), fixed a lot of pylint issues, add support to set the
top level log directory, or pass the log directory explicitly
(i.e. do not generate the log directory name). Also add an option
to skip creating the tar archive.

Signed-off-by: Hans Verkuil <hverkuil@xs4all.nl>