test_display.py: fix cec-stress-sleep test
Tests for the cec-stress-sleep command were missing, so it didn't work
at all.

Added this.

Moved the test for root permissions to the top, before it tries to wake
up the display. That way you get informed immediately if it isn't run
with sudo.

If -a is used with cec-stress-sleep, and max-sleep is missing, then add
max-sleep=5.