edid-decode: hide error when using -P and -a
When reading the EDID from the i2c device hide errors
if -P is also set. In that case just show f.f.f.f.

Do not hide if the EDID file does not exist, or if the
i2c device does not exist: those are too serious to hide.

Signed-off-by: Hans Verkuil <hverkuil@xs4all.nl>