edid-decode: support HF-EEODB when reading EDID over DDC
The code to read an EDID directly over the DDC bus did not support the
HDMI Forum EDID Extension Override Data Block, so EDIDs that support
that would ignore the extra extension data blocks.

Check for the presence of this data block and obtain the correct number
of extension blocks.

Signed-off-by: Hans Verkuil <hverkuil@xs4all.nl>