v4l2-ctl: process events before queues in stateful_m2m()
Events can change the streaming state and affect our ability (or
willingness) to further dequeue/requeue buffers.

For instance, a V4L2_EVENT_EOS means that we will stop streaming, so we
shouldn't process the OUTPUT queue any further once we receive it.
However, the current code processes the queues before the events, and
this can lead to issues such as an OUTPUT buffer dequeue attempt being
performed even though there is no buffer currently queued, resulting in
an indefinitely blocking DQBUF.

The issue was difficult to spot because (I assume) this usecase has been
mostly tested with software drivers like vicodec, which publish their
side-effects (notably in terms of events) before the ioctl that caused
them returns. But as far as I can tell this is not a requirement, and
events can also be sent asynchronously. In this case, a race condition
can cause the decoding loop to iterate one extra time and perform the
unfortunate DQBUF.

The fix is simple: process events before queues, so that the latter are
immediately affected by the result of published events instead of on the
next iteration of the loop. Events are supposed to be higher priority
anyway, and this is already done in that order in streaming_set_cap(),
so this sounds reasonable to do in any case.

Signed-off-by: Alexandre Courbot <gnurou@gmail.com>
Signed-off-by: Hans Verkuil <hverkuil@xs4all.nl>