cec-compliance: don't check UID, instead test with a <Poll>
The use of CEC_MSG_FL_RAW is only allowed if CAP_SYS_RAWIO is
set. So don't check if you are root, instead just try to transmit
a <Poll> message with CEC_MSG_FL_RAW set. If CEC_TRANSMIT returns EPERM,
then that means CAP_SYS_RAWIO was not available, and in that case return
NOTAPPLICABLE.

Signed-off-by: Hans Verkuil <hverkuil@xs4all.nl>