cec-ctl: fail if HPD goes high unexpectedly
In init_standby_wakeup_cycle_test() the display is put into standby.
Some displays pull the HPD low when in standby. This function checks
if that happened 5 seconds after the display reported that it is in
standby mode.

Wait two more seconds in that case and check if the HPD is still low
(i.e., the physical address is invalid). If it isn't, then something
weird is happening since you would not expect the HPD to go low and high
again while in standby.

Report a fail if this is detected.

Signed-off-by: Hans Verkuil <hverkuil@xs4all.nl>