test-media: add initial vidtv streaming test
Add a test to check that streaming I/O works for vidtv.

It would be nice if we could also compare the TS data, but this
contains timestamps, so it is always different each time you run
it. It would be nice if vidtv would have a module option to
enforce fixed timestamps.

Signed-off-by: Hans Verkuil <hverkuil-cisco@kernel.org>
Acked-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>