ir-ctl: remove quirky -rmw command line parsing
The --receive option takes an optional file name, even when specified as
a short option without an `='. So, ir-ctl -rmw saves the output to a file
called mw. You would expect the -r, -m, and -w options to be specified.

This commit removes the file argument completely. Saving to a file is
confusing and does not really have a good use-case. I've never heard of
anyone using this feature.

On top of that, I'm implementing a ir-ctl compatibility mode in cir[1],
and clap does not support short options with an optional value without
`='[2].

[1] https://gitlab.freedesktop.org/linux-media/tools/cir
[2] https://docs.rs/clap/latest/clap/struct.Arg.html#method.default_missing_value

Signed-off-by: Sean Young <sean@mess.org>