lib/libdvbv5/meson.build: use `includedir` also for `libdvb-version.h`
Header files are copied into the `includedir` by Meson.
However, `libdvb-version.h` is not treated as normal
include file, but copied with `configure_file`.
This means we have to explicitly declare the target directory.

The commit at hand fixes the target
directory to point to Meson's `includedir`.
That is necessary so the file ends up where all header files go,
even when the build outputs are split into separate packages.

Signed-off-by: Yarny <Yarny@public-files.de>
Signed-off-by: Hans Verkuil <hverkuil+cisco@kernel.org>