test-media: speed up stateful decoder compliance tests
The stateful decoder compliance tests used -s24 in combination with
--stream-from. But since the size of the output buffer is larger than
the provided file, that file is repeated 24 times. The file has 24
compressed frames, so all that you need to do is load the file in
the buffer and output it once, and you'll get 24 frames back.

So replace -s24 by -s1 for all these tests.

Signed-off-by: Hans Verkuil <hverkuil@xs4all.nl>