cec-compliance: fix Vendor Command With ID vivid tests
The vivid driver handled <Vendor Command With ID> incorrectly
if an unknown Vendor ID was used. This is now fixed, but as a
result the adapter tests in cec-compliance also have to be fixed.

According to the spec, if the given Vendor ID is unknown, then,
if it is a broadcast message, the command must be ignored, and
if it is a directed message, then it should be Feature Aborted.

This patch adds a test for the case of a broadcast message, and
fixes the test for a directed message, checking for Feature Abort
instead of expecting it to be ignored.

Signed-off-by: Hans Verkuil <hverkuil+cisco@kernel.org>