v4l2-ctl: do not fail on kernel not implementing VIDIOC_SUBDEV_QUERYCAP
The comment above the code snippet specifies that we do not want to fail
on kernels not implementing VIDIOC_SUBDEV_QUERYCAP (5.9 and earlier) but
calling doioctl will still set app_result to a negative value if the
kernel doesn't implement it (this ioctl returning -1 with errno set to
ENOTTY). Considering app_result is used as the exit code of the program,
ignoring the return value of doioctl function is not enough to
make this ioctl truly optional.

Let's handle this ioctl as optional to avoid kernels lacking it making
v4l2-ctl command fail (but still doing what it's been asked to do) by
calling test_ioctl directly.

Fixes: fc89f67d9fe0 ("v4l2-ctl: add support for VIDIOC_SUBDEV_QUERYCAP")
Signed-off-by: Quentin Schulz <quentin.schulz@cherry.de>
Signed-off-by: Hans Verkuil <hverkuil+cisco@kernel.org>