cec-ctl: warn if the SFT is too short
Warn if the detected Signal Free Time is too short. This helps
debugging SFT-related issues.

The SIGNAL_FREE_TIME defines are moved to cec-info.h so they can
be shared between cec-ctl and cec-compliance.

This patch accidentally ended up in a development branch, squashed
with an unrelated patch. But this is a useful addition to have when
debugging low-level CEC issues.

Signed-off-by: Hans Verkuil <hverkuil+cisco@kernel.org>