cec-ctl: incorrect LA issue with random stress test
The --stress-test-random-standby-wakeup-cycle sequence detects at the
start whether the display pulls the HPD low when in Standby, and
uses that information to determine whether Image View On should use
Unregisters as the initiator or not.

But some devices take some time before pulling HPD low when in Standby,
so depending on the timing, using Unregistered while the HPD is still
high will cause an error and the message was never transmitted.

Change the code to always check the current logical address before
transmitting Image View On.

Signed-off-by: Hans Verkuil <hverkuil+cisco@kernel.org>