v4l2-ctl: Fix failure on controls starting with a digit
Specifically querying the 3a_lock control from v4l2-ctl fails :

  $ v4l2-ctl -C 3a_lock
  VIDIOC_G_EXT_CTRLS: failed: Invalid argument
  : 0

A similar issue occurs when setting the control :

  $ v4l2-ctl -c 3a_lock=0
  VIDIOC_S_EXT_CTRLS: failed: Invalid argument
  Error setting controls: Invalid argument

This does not happen with other controls. Internally, v4l2-ctl parses
its arguments in order and differentiates a control name from a value
using this check :

  if (isdigit(s[0]))

This means if the argument starts with a letter, it is a control
name; otherwise, it is a control value. This works well for all controls
except for 3a_lock, which starts with a digit and is wrongly detected as
a control value.

Currently, 3a_lock is the only control that starts with a digit.

Introduce is_valid_number() to check if a string contains only digits,
and therefore would return false when parsing '3a_lock'. Use
is_valid_number() for control get and set operations.

Signed-off-by: Benjamin Mugnier <benjamin.mugnier@foss.st.com>
Signed-off-by: Hans Verkuil <hverkuil+cisco@kernel.org>